requirements.txt
streamlit
streamlit-ace
ply
requests
joblib
scikit-learn
numpy
colorama
rich
python-dotenv
